menu_service/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
aiokafka==0.12.0
sqlalchemy
python-dotenv
requests
python-multipart
pydantic
pydantic-settings